fenced-frame/content-shared-storage-get.https.html

<!DOCTYPE html>
<title>Test sharedStorage.get() in a fenced frame with network revocation.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>

// async function runSharedStorageTest(testcase) {
  // return testcase().finally(async () => {
  //   // Clear shared storage of same-site origin.
  //   await sharedStorage.clear();

  //   // Clear shared storage of cross-site origin.
  //   cleanup_frame = attachIFrameContext({origin: get_host_info().HTTPS_REMOTE_ORIGIN});
  //   await cleanup_frame.execute(async() => {
  //       await sharedStorage.clear();
  //   });
  // });
// }

promise_setup(async () => {
  // Set sharedStorage value for HTTPS_SAMESITE_ORIGIN
  await sharedStorage.set('test', 'apple');

  // Set sharedStorage value for HTTPS_REMOTE_ORIGIN.
  init_iframe = await attachIFrameContext(
                  {origin: get_host_info().HTTPS_REMOTE_ORIGIN});
  await init_iframe.execute(async () => {
      await sharedStorage.set('test', 'banana');
  });
});

promise_test(async (t) => {
  const fencedframe = await attachFencedFrameContext(
    {origin: get_host_info().HTTPS_SAMESITE_ORIGIN});

  await fencedframe.execute(async () => {
    await window.fence.disableUntrustedNetwork();
    let get_result = await sharedStorage.get('test');
    assert_equals(get_result, 'apple');
  });
}, 'Test sharedStorage.get() succeeds in a fenced frame with network revocation.');

promise_test(async (t) => {
  const fencedframe = await attachFencedFrameContext(
    {origin: get_host_info().HTTPS_SAMESITE_ORIGIN});

  await fencedframe.execute(async () => {
    try {
      await sharedStorage.get('test');
      assert_unreached('Calling get() without revoking network should throw.');
    } catch (e) {
      assert_equals(e.name, 'OperationError');
      assert_equals(e.message, 'sharedStorage.get() not allowed in fenced frames without first calling window.fence.disableUntrustedNetwork().')
    }
  });
}, 'Test that sharedStorage.get() in a fenced frame rejects when network is not revoked.');

promise_test(async (t) => {
  const fencedframe = await attachFencedFrameContext(
    {origin: get_host_info().HTTPS_REMOTE_ORIGIN});

  await fencedframe.execute(async () => {
    await window.fence.disableUntrustedNetwork();
    let get_result = await sharedStorage.get('test');
    assert_equals(get_result, 'banana');
  });
}, 'Test that sharedStorage.get() can only fetch keys for its own origin when network is revoked.');

promise_test(async (t) => {
    // First, try get() in a top-level document.
    try {
      await sharedStorage.get('test');
      assert_unreached('Call to get() in top-level document should throw.');
    } catch (e) {
      assert_equals(e.name, 'OperationError');
      assert_equals(e.message, 'Cannot call get() outside of a fenced frame.')
    }

    const samesite_iframe = await attachIFrameContext(
      {origin: get_host_info().HTTPS_SAMESITE_ORIGIN});

    // Then, try get() in an iframe.
    await samesite_iframe.execute(async () => {
      try {
        await sharedStorage.get('test');
        assert_unreached('Call to get() in top-level document should throw.');
      } catch (e) {
        assert_equals(e.name, 'OperationError');
        assert_equals(e.message, 'Cannot call get() outside of a fenced frame.')
      }
    });
}, 'Test that sharedStorage.get() rejects outside of a fenced frame.');

</script>
</body>
